Introduce abstractions for VAOs and VBOs, using them to build an initial, crude triangle abstraction
VAO/VBO resources are implemented in terms of a generic vertex_resource. This is templated both on the number of buffers and a type allowing us to customize generation/destruction. Under the bonnet, the latter is managed by vertex_resource_lifecyle, meaning that clients need only specify the bare minimum for representing a particular concrete resource.