Implement and test image loading
1. Create an image class for robustly loading and unloading images using the stb library. The class holds both the raw resource provided by stb and the width, height and number of channels. Under the assumption that stb correctly sets these values (i.e. its postconditions hold) it is an invariant of the class that the pointer manages width*height*channels elements.
a. Customize unique_ptr so that it frees images upon deletion.
b. Introduce a private 'parameter' type furnished with the appropriate move semantics to ensure that, even in the moved-from state, instances of image satisfy their invariants.
c. Use std::span to cleanly expose the underlying data to clients (some of which - notably glad - will require the non-const raw pointer).

2. Write false negative and standard tests. Note that as written with the exception testing at the bottom, ImageTest.cpp is sensitive to potential stb threading issues if run in parallel with the diagnostics tests. In a future lecture, we will ensure that such sensitivity is by design and not down to luck.