Err on the side of caution with noexcept
The noexcept on the first size is simply a copy-paste mistake.

If the class invariants are satisfied, the other two size() calls should never throw. But since they call functions which aren't noexcept, I'll be conservative and remove the noexcept.

This is quite an interesting design decision and I may change my mind!