Render our first texture!
1. shader_program enhanced so that int uniforms may be set.

2. The texture_2d_configurator has been expanded to allow things like the various filters and other parameters to be set as the last phase of configuration

3. A wrapper created for texture units

4. texture_2d furnished with a bind method, which activates the desired texture unit

5. polygon_base and derived classes enhanced to deal with arbitrary attributes. The presence of a texture is determined by detecting texture_coordinates<T> amognst the attributes. Several issues are deferred to the future:
a. Constraining the Attributes, which must tightly wrap fundamental types;
b. Dealing with multiple textures

6. The hexagon is now textured

This finally give rise to the first piece of textured geometry. However, as alluded to in kecture 25 there is a very subtle bug, and so the location is wrong for gcc/MSVC (but correct for clang). This will be resolved next time...