Switch to a memory-ordered tuple and add type constraints
The reason the textured hexagon was rendered incorrectly on MSVC and gcc - but not clang - was because it is only in libc++ that the elements of std::tuple are naturally ordered. Switching to sequoia::mem_ordered_tuple gives the correct results across platforms.

In addition to this change, traits have been added to check whether a type is legal for use in a OpenGL buffer. Ultimately, any such type must tightly wrap one of OpenGL's arithmetic types.

There are two components to this:
1. template<class T> gl_arithmetic_type_of; which exposes a type named type if and only if
a. Either T is an arithmetic type
b. Or T recursively exposes value_type with the recursion terminating in an arithmetic type.

2. template<class T> is_legal_gl_buffer_value_type; This provides a mechanism for clients of buffers to opt their types in. We do this for mem-ordered_tuples under certain conditions and also for sequoia::maths::vec_coords, as long as the latter isn't padded. Reflection should offer a superior approach.