Simplifications, since getting the OpenGL version is cheap
With glad updated to core-4.6 (which is what I should have used from the start), there is now a cheap way to get the supported OpenGL version (see Version.cpp). Since it is no longer the case that this involves a OpenGL call, gl_function may be simplified without incurring the cost of an unwanted OpenGL call.

Actually, a static local variable was used to ensure that the call only ever happened once, so wasn't really worth worrying about. That being said, the presence of the static  was not very desirable and so the new code is cleaner and simpler.