Protect against abuse of gl_function
Ordinarily, gl_function should be constructed just with the desired function pointer. However, when setting up the OpenGL advanced error checking, it is necessary to suppress error checking, else - in the case of an error occurring- we end up the unfortunate situation of trying to use advanced error handling while in the process of setting it up! Opting out can be done conveniently by using the constructor overload which takes unchecked_debug_output_t as the first parameter, A deduction guide ensures that gl_function's second template parameter - the debugging_mode - set to debugging_mode::none.

Previously, it was conceivable that someone could abuse this by explicitly setting the Mode to something other than 'none' and then calling the unchecked_debug_output_t constructor overload. If this is attempted, clients will now encounter a hard error, courtesy of the newly-added static_assert.