Pre-supply two simple classes for lecture 29
contextual_resource_handle wraps both a context and a resource_handle. The latter has always had move semantics such that move assignment performs a swap. This is to ensure that GPU resources are properly cleaned up upon destruction of moved-from state.

To ensure that contextual_resource_handle has consistent move semantics, the context is wrapped inside context_ref, the sole purpose of which is to provide the correct move semantics.