Get the Demo project compiling (and working!) now that the glad backend has been swapped out
This is largely a matter of propagating the context created by the window to all the places where there is an OpenGL call.

To fit this into a (slightly extended) lecture, a few shortcuts have been taken.
1. Some naming can be improved
2. destroy functions more naturally take a contextual_resource_handle than separate context and raw_indices
3. The contextual_resrouce_handles suffers from being an array of N constextual_resrouce_handles, rather than a single context reference and N resource_handles. This is the origin of the rather janky context() method. Fixing this will be the subject of a later lecture.

Note that this commit leaves the test project broken, which will be fixed shortly.