Minimize the amount of code within the scope of preprocessor directives
As we saw during the live coding, having code which is removed by the preprocessor is not great. The code needed to be changed during refactoring, but the IDE and compiler were not able to give any clues as to whether we'd done it correctly. This is a signal to move as much of this code as possible to a place where it is always visible to the compiler.

1. Tidy up the preprocessor logic at the top of the file.
2. Move the clang-specific code out from inside the preprocessor directives.
3. Make use of a 'feature test macro' which allows us to more precisely state the intent of the preprocessor sections containing code.