Build infrastructure for customizable error handling
Customizable error handling is recognized as a special case of python-like decorators, meaning that each dispatch done to the GPU by gl_function comes with a prologue and an epilogue. To do this, we exploit the fact that a GladGLContext is now propagated to all OpenGL call sites.  The glad context is wrapped inside a decorated_context, together with the debugging mode and prologue/epilogue function objects.

The new structure allows simplifications to gl_function, which no longer needs to infer the debugging mode. However, it is still necessary to provide an escape hatch to override the context's debugging mode, temporarily forcing it off. As before, this is to avoid problems in two edge cases:

1. When setting up advanced debugging, it is necessary to make OpenGL calls; but these had better not check for errors since the debugging environment hasn't been set up!

2. When extracting error messages, the OpenGL calls involved shouldn't themselves try to extract error messages since this results in infinite recursion.

Note that to achieve this the old tag struct unchecked_debug_output_t is replaced with the more idiomatic debugging_mode_off_type using standard std::integral_constant tricks.

The existing error checking behaviour is preserved by installing default_error_checker as the default epilogue in window_config. This is at our highest architectural level, and we have avoided setting defaults lower down the chain, to force clients to make an active choice of behaviour. Strictly, the existing error checking is actually improved, since we now also report the name of the offending OpenGL function that causes the issue.