Experiment with stenciling and start prototyping an OpenGL capability manager
The capability_manager is just a sketch.

For example, it captures the context by const&, which doesn't play nicely with copies/moves. However, the semantics for this type are subtle and we will deal with them properly in future.

For now, the goal is to get the basic mechanics working and to test them, to ensure that we're roughly on the right tracks.

As such, state management for multisampling and blending has been introduced. Note, though, that the treatment for the latter is currently incomplete since glBlendColour is neglected.

Some testing has been done, but the test has a linear structure making it hard to audit what is actually being tested. Similarly, it is awkward to add new checks that start from the desired state but leave the state after completion as we intended for anything that follows. Next time, we'll see a vastly superior way of building such tests.