Adopt a consistent convention for member functions which are const but have side effects
Since early in the course, we have agonized about how to deal with member functions that are const but have side-effects. In the context of the rendering engine, the latter are most generally calls to the GPU.

While not incorrect to simply mark such member functions const, this perhaps has historical baggage that may imply to some that the member function is a pure function. Earlier in the course, I offered friend functions as a convention to express that while the object parameter is const&, there's no reason to assume that the function doesn't mutate other state, unless documentation says so.

As the course has evolved, we've started making use of explicit object parameters (deducing this). To my mind, this also (amongst other things) expresses the same convention. Although the meaning is the same as a const member function, I feel that making the object parameter explicit is a form of demystification and I think plausibly telegraphs that no assumption can be made about potential side effects, unless the function is documented as such.

Therefore, in this commit I retire the friend convention and utilize explicit object parameters to harmonize our approach.