Continue to migrate code from CapabilityManagerFreeTest.cpp into avocet
1. Move the code which configures capabilities into a new .hpp/.cpp. Note that part of the motiviation is to mitigate a cyclic dependency which may otherwise arise if we later wish to incorporate the capability_manager itself into the decorated_context. Therefore, CapabilitiesConfiguration.hpp forward declares decorated_context and contains only declarations (except in trivial cases). The definitions - which require gl_function (and hence the definition of the decorated_context) reside in the cpp.

2. The to_string functions for the various capability-related enums now have just declarations in Capabilities.hpp, with the definitions in the associated .cpp

In both cases above, we prefer reducing dependencies over the possible optimization benefits of keeping code in headers. Any performance gains are expected to be negligible. If evidence arises to the contrary, we can review our decisions. However, quite apart from using hpp/cpp splitting to mitigate potential cyclic dependencies, our preference is generally for reduced dependencies and improved compilation times. It is quite conceivable that we may wish to change the precise stringification in to_string; done in the hpp this will cause pretty much everything to recompile.