Significantly reduce test execution time on windows
Rather than creating a gflw_manager per test, the graphics tests come with a static instance. This reduces the test execution time with MSVC (debug/asan) by over a factor of 2.

Note: great care must be taken her. glfw_manager does not work properly if instances have overlapping lifetimes. This can happen if the glfw_manager is a static member of basic_graphics_test: it gets created before the instance used to determine the rendering setup and everything breaks.

There are a number of obvious extensions:
1. Make the system less brittle
2. Enable parallel execution of graphics tests. This can probably be done by ensuring windows are created at the point of construction and then made current at the start of run_tests ( which will be executing off the main thread).