Fix for (linux) systems with OpenGL version >= 4.3
The problem is with the new consistency check in decorated_context_base::init_debug. If the debuggin_mode is not off, and debug output is supported, then the expectation is that the context is a debug context. This is fair enough. The problem comes when trying to divine the OpenGL version. For linux we start with a trial version of 4.1. This means that the window hit for debugging is false. However, when the window is actually created, the OpenGL version turns out to be (say) 4.5 which does support debugging. This gives an apparent conflict. The simplest solution is not to request a debug context when trying to determine the support OpenGL version.