Separate out functionality from decorated_context into context_base and contexty
Primarily, this commit simplifies code along the lines of the lecture 35 presentation. The inspiration comes from the fact that, previously, we had this somewhat awkward method to invoke gl_function with the debugging_mode temporarily treated as-if off. There were two scenarios where this was necessary:
1. When setting up advanced debugging, since one cannot check for errors while the mechanism for doing so is being enabled/configured;
2. When making the final call to actually get errors (either advanced or basic), since checking for errors caused by getting errors leads to infinite recursion.

We now provide a nicer way of doing this, albeit with a tradeoff.
1. context_base now performs the initialization of the debug environment.

2. context inherits from this and additional stores per-context information that we only want to extract from the driver once. Currently, this is just the maximum length of debug messages, but in a separate commit we'll subsume much of what is in Version.hpp

3. decorated_context inherits form context and provides the prologue /epilogue.

The point is that, for example, context_base::init_debug knows nothing about prologues and epilogues; there's no way to do error checking will in the function. Similarly, we choose to change the signatures of things in Errors.hpp and Version.hpp so that they take either context_base or context: either way they no longer take decorated_context and so
1. The dependencies simplify
2. The need for manually ignoring the debugging_mode disappears.

This leads to a nice simplification of gl_function, since code involving debugging_mode just disappears.

The only downside is that some of our OpenGL function calls now can't come with any prologue/epilogue. Therefore, if we want to use, say, the epilogue to log every OpenGL function call, we can't. However, the only ones we'll miss are either
1. Calls we'll make once when a context is being created, to determine things like the maximum debug message length
2. The final call made as part of debugging to actually extract any debug messages from the driver.

Overall, the benefits of the simplifications outweigh the downsides, not least since we'll still be able to log almost everything of interest.

Additionally, I have taken the opportunity in this commit to create some utilities for formatting and have partially rolled them out across the code base. This will be completed in a subsequent commit.