Improve platform independence
This commit is a nice example of why it's a good idea to split these sorts of functions between an hpp and a cpp. Had this been in the hpp, the entire project would have to be recompiled after this change.

Of course, in some cases it may be desirable for a function to end up in the header to give the optimizer inlining opportunities. That is unlikely to be the case for this function and, even if it is, it can be migrated once we have high confidence that it's not going to be changed.