Integrate capability management into avocet
To facilitate this, we start by adding a new class to the tip of context hierarchy. capable_context inherits from decorated_context, and so the latter now has protected destructor/moves. To proceed, we could make an instance of the capability_manager a private member. However, there are some hoops to jump through:

1. We'd need to either make capability_maager a private type or make its constructor private but grant capable_context friendship. This would prevent there being more than one capability manager per context;

2. As it stands, the capability_manager constructor takes a decoarated_context const&. If we preserve this, we'd end up with a self-referential type. This is not a disaster, but we'd have to take care implementing moves.

Rather than going down this route, we just pull code out of capability_manager into capable_context and then junk the capability_manager.

Now to the most controversial piece of design: I have made capable_context::m_Payload mutable. I very rarely use this technique, but there's a place for everything, Here is the reasoning:

The capability management process is a black-box. It so happens to elide unnecessary OpenGL calls, but that is an implementation detail, opaque to clients. So suppose for a moment that it did not do this. then m_Payload would cease to exist. In this scenario capable_context is similar to other classes we have: it has member functions which preserve CPU state but have side effects on the GPU. We have been using a convention to denote such functions: explicit object parameters, qualified with const&. This makes the function signatures more consistent with free functions with the interpretation being: just because a function has const parameters, it can't be assumed not to have side effects unless the documentation says so.

So, imagine this is how we implemented things. Then we decide, purely as an optimization, to implement opaque state tracking. Should the function signatures change? I think there is a good argument to leave them as-is, because of the black-box nature of the optimization. The way to achieve this is with the mutable keyword.

It is important to note that this approach does not create new multi-threading issues: The same context may not be current on two threads and it's already UB according to OpenGL if this happens. Besides which, setting payloads will always have side effects albeit on the GPU. We're adding CPU side effects, but I think in a controlled and deliberate way. I therefore think this a a rare example of a compelling use of mutable. However, I could be convinced otherwise!

Some other benefits we get from this is that we are able to preserve the const propagation of contexts through our code. This is important since it prevents clients from, say, reassigning a context with one that has nothing but null function pointers. It is desirable to preserve this. Of course, one could always do this while, say, reinstating the capability_manager as a member and granting mutable access to this, but considerable care would have to be taken to prevent the wrong clients mutating through assignment, while allowing things like the window manager to perform moves.

Overall, it seems like mutable is the right tool for this particular job...