Start adding support for depth testing
This commit introduces the necessary structs and gpu configuration calls.

Note that the gl_depth_test includes not only the 'func' and 'mask' but also the polygon offset. The latter it utilized by three other capabilities, GL_POLYGON_OFFSET_* and in each case affects depth testing. Therefore, it makes sense for the state to reside with the depth test itself.

Integration into the existing capability management system, and testing, will be done in  a subsequent commit.